--- a/packages/list-reusable-blocks/package.json
+++ b/packages/list-reusable-blocks/package.json
@@ -22,6 +22,7 @@
 	"engines": {
 		"node": ">=12"
 	},
+	"gutenberg:source": "src/index.js",
 	"main": "build/index.js",
 	"module": "build-module/index.js",
 	"dependencies": {