--- a/packages/media-utils/package.json
+++ b/packages/media-utils/package.json
@@ -22,6 +22,7 @@
 	"engines": {
 		"node": ">=12"
 	},
+	"gutenberg:source": "src/index.js",
 	"main": "build/index.js",
 	"module": "build-module/index.js",
 	"dependencies": {