--- a/webpack.config.js
+++ b/webpack.config.js
@@ -4,5 +4,8 @@
 const blocksConfig = require( './tools/webpack/blocks' );
 const developmentConfigs = require( './tools/webpack/development' );
 const packagesConfig = require( './tools/webpack/packages' );
+const experimentalConfig = require( './tools/webpack/experimental.config' );
 
-module.exports = [ blocksConfig, packagesConfig, ...developmentConfigs ];
+module.exports = process.env.EXPERIMENTAL_BUILD
+	? experimentalConfig
+	: [ blocksConfig, packagesConfig, ...developmentConfigs ];