<!-- Learn the overall process and best practices for pull requests at https://github.com/WordPress/gutenberg/blob/HEAD/docs/contributors/repository-management.md#pull-requests. -->

<!-- Gutenberg's license is in the process of updating to be dual-licensed under the GPL and MPL. As part of that transition, all new contributions are dual-licensed. For more information, see: https://github.com/WordPress/gutenberg/blob/trunk/LICENSE.md -->

## Description
<!-- Please describe what you have changed or added -->
Part of https://github.com/WordPress/gutenberg/discussions/36142.

Currently, our development workflow works roughly like this:
1. `npm run dev`: which runs (2), (3) and (4).
2. `npm run build:packages`: build all packages with Babel for js/ts and SASS for css into _package-level_ build directories (`build`, `build-module`, `build-style`) once.
3. `npm run dev:packages`: start watchers to watch for all the source files and rebuild (2) when they change.
4. `wp-scripts start`: run webpack watch mode and watch for the **build** files built by (2) and (3) then bundle them to _root-level_ `build/` directory.
5. Browsers import files in (4).

The problem is that whenever we change something in our source code, we have to wait for Babel/SASS to rebuild them into _package-level_ builds (3), and then wait until webpack pick them up and bundle them into the final build. It's slow for both initial build and incremental build, since that a lot of expensive IO are wasted on building intermediate results. While we do need the intermediate _package-level_ builds for production and deployment, we don't need them during development.

This PR tries to simplify that by leveraging webpack to do all the heavy lifting during development. It'll handle all the JS files along with all the CSS files, directly from their sources. We won't be outputting intermediate builds, no custom build scripts and extra watchers. The build performance should be a lot faster for initial build and more responsive for incremental builds too.

This PR introduces a new command `npm run dev@exp` to run the experimental `dev` build. It runs on a different path than the original `dev` command behind a feature flag, so that we can safely test it without risking breaking everyone's build.

I ran a couple of builds locally to compare between the two, and below is what I saw on my macbook pro 2020 (2.3GHz Intel Core i7, 16GB RAM, macOS 12.0.1). (Since that both commands will start watchers and never finish, we have to do some some tweaks to be able to run the `time` audit. First, go to `packages/scripts/scripts/start.js` and remove `'--watch'`. Then, for `npm run dev`, you'll have to remove `npm run dev:packages` in the scripts in `package.json`. Finally, for `npm run dev@exp`, you'll have to comment out `pollTimeout: ...` in `tools/webpack/experimental.config.js`) (You might want to run each the second time to take cached/warmed results into account.)

**For initial build:**

```
# time npm run dev
225.23s user 23.39s system 339% cpu 1:13.22 total

# time npm run dev@exp
48.61s user 5.65s system 185% cpu 29.255 total
```

The total amount of time is reduced to about **40%** and the CPU usage is only **55%** too. 

**For incremental builds:**

I don't have a consistent setup for auditing incremental build performance though, so below is just a rough number in my local testing:

`dev` | `dev@exp`
--- | ---
~3s | ~1.5s

It's still only around **50%** of the incremental build time though.

Note that this change isn't only for performance, but it also makes implementing other enhancements in https://github.com/WordPress/gutenberg/discussions/36142 easier in the future, performance boost is just a bonus.

## How has this been tested?
<!-- Please describe in detail how you tested your changes. -->
<!-- Include details of your testing environment, tests ran to see how -->
<!-- your change affects other areas of the code, etc. -->
1. `npm run dev` should work the same
2. `npm run dev@exp` should also work as expected
3. Do a fresh clone of the repo and run `npm install` and `npm run dev@exp` should also work

## Screenshots <!-- if applicable -->

![image](https://user-images.githubusercontent.com/7753001/144191850-af8f4e14-0f4b-4338-ab66-a00a1d8f5a34.png)


## Types of changes
<!-- What types of changes does your code introduce?  -->
<!-- Bug fix (non-breaking change which fixes an issue) -->
<!-- New feature (non-breaking change which adds functionality) -->
<!-- Breaking change (fix or feature that would cause existing functionality to not work as expected) -->
New feature

## Checklist:
- [ ] My code is tested.
- [x] My code follows the WordPress code style. <!-- Check code: `npm run lint`, Guidelines: https://developer.wordpress.org/coding-standards/wordpress-coding-standards/javascript/ -->
- [ ] My code follows the accessibility standards. <!-- Guidelines: https://developer.wordpress.org/coding-standards/wordpress-coding-standards/accessibility/ -->
- [ ] I've tested my changes with keyboard and screen readers. <!-- Instructions: https://github.com/WordPress/gutenberg/blob/HEAD/docs/contributors/accessibility-testing.md -->
- [x] My code has proper inline documentation. <!-- Guidelines: https://developer.wordpress.org/coding-standards/inline-documentation-standards/javascript/ -->
- [ ] I've included developer documentation if appropriate. <!-- Handbook: https://developer.wordpress.org/block-editor/ -->
- [ ] I've updated all React Native files affected by any refactorings/renamings in this PR (please manually search all `*.native.js` files for terms that need renaming or removal). <!-- React Native mobile Gutenberg guidelines: https://github.com/WordPress/gutenberg/blob/HEAD/docs/contributors/code/native-mobile.md -->