@adamziel 2021-11-22
I wonder if this could replace webpack in the longer run: https://swc.rs/