@kevin940726 2021-11-22
> I wonder if this could replace webpack entirely in the longer run: https://swc.rs/

That could replace Babel, it's mentioned in the original discussion in #36142.