@noisysocks 2021-11-22
> Size Change: -240 kB (-18%) 👏

Wasn't expecting this. Did it mess up the Size Change script?