@kevin940726 2021-11-23
> Wasn't expecting this. Did it mess up the Size Change script?

Fixed!