@gziolo 2021-11-23
Thank you for starting this PR. It looks like the direction in the JavaScript community is to further decouple the development process from production builds. Vite is the best example for that. 

> **For initial build:**
> 
> ```
> # time npm run dev
> 152.03s user 15.98s system 435% cpu 38.564 total
> 
> # time npm run dev@exp
> 27.25s user 3.18s system 148% cpu 20.496 total
> ```
> 
> The total amount of time is reduced to nearly **50%** and the CPU usage is only **1/3** too. 

`npm run dev` includes also additional steps to build and verify types with TypeScript which might have significant impact on the comparison:

https://github.com/WordPress/gutenberg/blob/88883a3ea23f5b3d8aff563f6a8447e4133b9e55/package.json#L228

It would be great to measure it. Also, are we sure we don’t need want to bother with types during development? I’m wondering what the experience would be in case someone works on changes to the packages that require modifications to TypeScript definitions.