@kevin940726 2021-12-01
@gziolo Thank you for mentioning that! I added `fork-ts-checker-webpack-plugin` to help do type checking during development. This is also a new feature since that I don't think we're checking TS types for incremental builds currently.