@noisysocks 2021-12-02
When testing locally on an already very fast computer I see a ~2x speed up (15s → 7s) for initial build and a ~4x(!!) speed up (2s → 0.5s) for incremental build. Definitely will improve some lives!

I like that it's a seperate `npm run` command so that we can test this in real world conditions first to see if there are issues. I'd say if GitHub Actions is happy then let's give it a go.