@gziolo 2021-12-13
In https://github.com/WordPress/gutenberg/pull/36736, @youknowriad explores speeding up the build by splitting the tasks into individual packages and using SWC instead of Babel. It doesn't necessarily mean this is an alternative for this PR, but we should explore a wider set of options before we jump to the conclusion of how we should improve the development workflows.