@kevin940726 2021-12-14
> The approach in this PR has a risk of hidden bugs because of the different build pipelines between packages and plugin. I think if we can speed up the packages build and optimize the bundling step, it'd be an ideal approach.

Why? We're already building different builds between packages and plugin, just that the plugin build depends on the packages build, the final output is still different. Furthermore, this PR only targets development build, so the production build is still intact.

However, IMO plugin and packages build should still have different pipelines, as long as they are using the same source files. Which build pipeline they use, whether it is webpack, babel, or SWC, is implementation detail, and it shouldn't matter for the final results. It's even better as we can apply different optimizations for different pipelines. For instance, JSX transpilation can continue to use `createElement` for package builds, and use the new automatic runtime for plugin build. Another example is that webpack can parse dynamic imports differently and enable code splitting and lazy loading for those assets.

> The other thing I came up with is that we might not be able to assume that all packages are built in the exact same way in the future, and having the flexibility for each package to define its own way to build is a good thing (in addition to sharing common building using wp-scripts or similar)

I agree, the `gutenberg:source` field in `package.json` I added in this PR is an example of that. I don't think we need to specify all the details though, for most packages we only just need the source entry files. We can add some other fields for other special packages if necessary.