@youknowriad 2021-12-16
> Why? We're already building different builds between packages and plugin, just that the plugin build depends on the packages build, the final output is still different. Furthermore, this PR only targets development build, so the production build is still intact.

The main issue for me is that Core consumes the production build but I've been thinking more and I think it's not that important especially since our CI also uses the production build. So I wouldn't be against a separate pipeline. 

That said, I wonder if we might be able to get good optimizations for both production and dev tooling potentially if we optimize the way we build the deps (see tools like rush, turbo, nx) while avoiding the new build tool to maintain (dev specific one).

Anyway, I think our north star in terms of build tooling should be to avoid custom tooling and custom things in general to retain as much flexibility as we can and avoid locking. (custom loaders, babel plugins...) There's a lot of innovation in this space at the moment and we shouldn't prevent ourselves from trying different approaches.