@kevin940726 2021-12-16
> That said, I wonder if we might be able to get good optimizations for both production and dev tooling potentially if we optimize the way we build the deps (see tools like rush, turbo, nx) while avoiding the new build tool to maintain (dev specific one).

I agree! I actually tried hard to make this PR work for production too, but I haven't tested it to be sure. So in the future, hopefully we can optimize production builds with the help of this PR too.
 
> Anyway, I think our north star in terms of build tooling should be to avoid custom tooling and custom things in general to retain as much flexibility as we can and avoid locking. (custom loaders, babel plugins...) There's a lot of innovation in this space at the moment and we shouldn't prevent ourselves from trying different approaches.

That's also my goal with this PR! Most modern build tools nowadays expect a single-pass pipeline to optimize the performance. If this PR works, then I believe we just prove that we don't need special handling or multiple pipelines. By returning to the basic, common way of configuring our webpack configs, I believe we can also make the migration path smoother if we ever want to move to other build tools. webpack is still the most popular build tool for web apps right now, so that we can expect next-gen build tools to provide smoother migration path from it.

What do you say if we merge this PR and try it out? The whole thing is still behind a flag though so we can experiment more stuffs on it. I have other optimizations planned ahead (#36142), but we need this to be merged to be able to work on those stuffs 😅 .