@youknowriad 2021-12-16
>  So in the future, hopefully we can optimize production builds with the help of this PR too.

Actually, I think we should always stick with the two steps build for production, because that's how Core consumes the packages (it consumes the built packages) and we shouldn't derail from that otherwise we risk having bugs that only show up in Core but not Gutenberg.

> What do you say if we merge this PR and try it out? 

not against trying this but not sure we need the flag, everyone working in the project should experience it, we can always revert or whatever.